feat: refactor

+++ src/main/java/br/jomoliveira/citel/dtos/DoadoresPorReceptorDTO.java
@@ -1,4 +1,4 @@
 package br.jomoliveira.citel.dtos;
 
-public record DoadoresPorReceptorDTO(String sorotipagem, Long quantidade) {
+public record   DoadoresPorReceptorDTO(String sorotipagem, Double quantidade) {
 }
